Ensure we select the ideal service version to mutate during compute init.
